+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Version 0.2.3  -  16/11/2022
 
 - Made OnSceneLoad public and accessible 
+- Changed name of OnSceneLoad -> OnSceneCollectionLoaded
+- Added another event, now there is two. one which outputs the scene collection data and one void.
 
 # Version 0.2.2  -  14/11/2022